# Conflicts:
#	app/Category.php
#	app/Http/Controllers/ShopController.php
#	app/Http/Controllers/UserController.php
#	app/Order.php
#	app/OrderDetail.php
#	app/Product.php
#	app/User.php
#	public/bootstrap/js/common.js
#	resources/views/back-end/login.blade.php
#	resources/views/front-end/pages/login.blade.php
#	routes/web.php

+++ app/Http/Controllers/UserController.php
@@ -49,7 +49,9 @@ class UserController extends Controller
         $user->status = '1';
         $user->role = $request->role;
         $user->save();
-        return redirect()->route('users.create')->with('alert', 'Thêm mới thành công');
+
+        return redirect()->route('users.index')
+            ->with(['message' => 'Thêm mới thành công', 'alert' => 'success']);
     }
 
     /**
@@ -86,16 +88,12 @@ class UserController extends Controller
     public function update(UserRequest $request, $id)
     {
         $user = User::find($id);
-        $user->password = bcrypt($request->password);
-        $user->fullname = $request->name;
-        $user->birthday = $request->birthday;
-        $user->phone = $request->phone;
-        $user->address = $request->address;
-        $user->gender = $request->gender;
         $user->status = $request->status;
         $user->role = $request->role;
         $user->save();
-        return redirect()->route('users.edit', $id)->with('alert', 'Chỉnh sửa thành công');
+
+        return redirect()->route('users.index')
+            ->with(['message' => 'Chỉnh sửa thành công', 'alert' => 'success']);
     }
 
     /**
@@ -107,7 +105,17 @@ class UserController extends Controller
     public function destroy($id)
     {
         $user = User::find($id);
+        $products = $user->products;
+        if (count($products) > 0) {
+            return redirect()->route('users.index')
+                ->with([
+                    'message' => 'Không thể xóa tài khoản này! Tài khoản này đang có sản phẩm.',
+                    'alert' => 'danger'
+                ]);
+        }
         $user->delete();
-        return redirect()->back()->with('alert', 'Xóa thành công!');
+
+        return redirect()->route('users.index')
+            ->with(['message' => 'Xóa thành công', 'alert' => 'success']);
     }
 }
